# Precompiled headers and improved include structure