--- a/libshvbroker/src/appclioptions.cpp
+++ b/libshvbroker/src/appclioptions.cpp
@@ -1,4 +1,4 @@
-#include "appclioptions.h"
+#include <shv/broker/appclioptions.h>
 
 #include <shv/chainpack/irpcconnection.h>