--- a/libshvbroker/src/rpc/brokertcpserver.cpp
+++ b/libshvbroker/src/rpc/brokertcpserver.cpp
@@ -1,7 +1,7 @@
 #include "brokertcpserver.h"
 #include "ssl_common.h"
 #include "clientconnectiononbroker.h"
-#include "../brokerapp.h"
+#include <shv/broker/brokerapp.h>
 
 #include <shv/coreqt/log.h>
 #include <shv/iotqt/rpc/socket.h>