--- a/libshvchainpack/include/shv/chainpack/abstractstreamwriter.h
+++ b/libshvchainpack/include/shv/chainpack/abstractstreamwriter.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "rpcvalue.h"
-#include "../../c/ccpcp.h"
+#include <shv/chainpack/rpcvalue.h>
+#include "../../../c/ccpcp.h"
 
 #include <ostream>