--- a/libshvchainpack/src/chainpack/abstractstreamreader.cpp
+++ b/libshvchainpack/src/chainpack/abstractstreamreader.cpp
@@ -1,4 +1,4 @@
-#include "abstractstreamreader.h"
+#include <shv/chainpack/abstractstreamreader.h>
 
 namespace shv::chainpack {