I have removed the previous headers structure, as there's no need for it. Public headers (those that get installed) are in the `include/` directory, and private (helper files, that the consumer of the library doesn't use) stay in the `src/` directory.

I have enabled header precompilation, it seems to work out of the box and speeds up the build process quite a lot!
Here are the results of non-cached build on my 8c/16t machine
```
Before:
real	1m15.332s
user	13m8.326s
sys	0m56.782s

After:
real	0m50.024s
user	5m18.855s
sys	0m39.667s
```
This is roughly a 33% improvement, which is nice. ~~However, a cached full rebuild does a little worse, I'm not sure why. It could be that caching doesn't work all that well, if you're using precompiled headers. Although there is support (or at least ccache says so).~~ Nevermind, caching works, you just need to set these variables:
```
export CFLAGS=-Xclang -fno-pch-timestamp
export CXXFLAGS=-Xclang -fno-pch-timestamp
export CCACHE_SLOPPINESS=pch_defines,time_macros
```
 However, I feel like overall, compilations will be faster. Hopefully.

Let's see how the CI compilation speed increases.