@syyyr 2023-10-25
CI compilation results:
```
+--------------------------+---------+---------+---------+
| Workflow                 | Before  | After   | Change  |
+--------------------------+---------+---------+---------+
| Qt 5.15.2 / Ubuntu 22.04 | 15m 38s | 13m 21s | -14.60% |
| Qt 6.5.0  / Ubuntu 22.04 | 19m 10s | 13m 18s | -30.60% |
| Qt 6.5.0 / Windows       |  9m 38s | 10m 51s |  12.62% |
| Qt 6.5.0 / NixOS         |  4m 36s |  3m 23s | -26.44% |
+--------------------------+---------+---------+---------+
```
- all Linux jobs got a little faster
- the Windows jobs got a little slower - I'm not sure why that is, but it's not a problem because the Linux jobs are slower overall