@syyyr 2023-10-25
I have also changed the CI so that pushes to the same PR or to the same branch cancel previous runs of workflows.