@syyyr 2023-10-31
After some discussion, precompiled headers is probably not what we want. ccache might be, but I'm closing this anyway, to split the include/ stuff from this.